apps/api/pyproject.toml
[project]
name = "api"
version = "0.0.1"
description = "REST API for NSG tool."
readme = "README.md"
requires-python = ">=3.12.8"
dependencies = [
    "django>=5.1.5",
    "django-cors-headers>=4.6.0",
    "django-ninja>=1.3.0",
    "ninja-extended",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.7.1",
    "toml>=0.10.2",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.1",
    "pytest>=8.3.4",
    "pytest-django>=4.9.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.9.2",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings_test"

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py312"
exclude = ["manage.py", "**/migrations/*.py"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    #### modules
    "ANN", # flake8-annotations
    "COM", # flake8-commas
    "C90", # mccabe complexity
    "DJ",  # django
    "EXE", # flake8-executable
    "T10", # debugger
    "TID", # flake8-tidy-imports

    #### specific rules
    #"D100",   # ignore missing docs
    #"D101",
    #"D102",
    #"D103",
    #"D104",
    #"D105",
    #"D106",
    #"D107",
    #"D200",
    #"D205",
    "D202",
    "D203",
    "D211",
    "D212",
    "D213",
    #"D400",
    #"D401",
    "D413",
    #"D415",
    "E402",   # false positives for local imports
    "E501",   # line too long
    "ISC001",
    "S101",   # assert
    "TRY003", # external messages in exceptions are too verbose
    "TD002",
    "TD003",
    "FIX002", # too verbose descriptions of todos
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "DTZ", "PLR"]

[tool.uv.sources]
ninja-extended = { git = "https://github.com/lucas-camp-it-soultions/ninja-extended.git" }
